add offer player model to replace the HABTM join table, and add tests for the offer player model
I decided to use the join table and the has_and_belongs_to_many relationship because I did not think
that I would need any additional checks or logic in the relationship between the two. After realizing
that in-game trade offers are limited to four players, and after having difficulty making a clean form
to allow for adding multiple players to an offer at once, I decided the best solution was to create
a model to represent the join table, place additional logic in that model, and replace the HABTM relationships
with the has_many through relationships. Unlike other tables with foreign keys, the offer_players
model does not check the reference on the offer because I am using a nested form, where the offer
will be created at the same time as the offer_players.